✨ feat(server): sync substrate → commonMain (SqlSyncableRepository/ChangeBus/SyncRegistry; digest on Sha256 seam; atomicfu registry)
